chahua: 重构 renderer.js（一）—— 抽出 message_filter / commands 模块
renderer.js 是 1671 行的 god file。按既有「一功能一模块」的工厂模式分阶段拆分。

本阶段抽两个最干净的自包含单元（无共享可变态、纯回调依赖）：

- message_filter.js —— 消息任务 chip + filter 视图。filter 状态（_filterTaskId /
  banner DOM）全封进模块，renderer 只经 afterAppendMessage / exitFilter /
  onTaskStateChange 三个出口交互。
- commands.js —— 斜杠命令解析（/help /tools /skills /task /clear task /clear
  /new）。tryHandle(text) 命中即执行返 true。HELP_TEXT 随之迁入。

renderer.js 1671 → 1470 行。行为零变化，三文件 node --check 通过。
